Implement dependency checking internally
Instead of needing CPAN::Meta::Check, we do this by hand as we're
already iterating modules to find versions from @INC.

The test report header has been revised for clarity about showing all
modules listed, including optional ones and the warnings header has been
revise to clarify that it is only about required dependencies.

This also implements a fallback to META if MYMETA is not available.